+++ extensions/obsidian-bookmarks/src/views/FileListItem.tsx
@@ -1,6 +1,7 @@
-import { Icon, List } from "@raycast/api";
+import { List } from "@raycast/api";
 import { Dispatch, SetStateAction } from "react";
 import DetailsActions from "../actions/DetailsActions";
+import getFaviconIcon from "../helpers/get-favicon-icon";
 import { File } from "../types";
 import FileItemDetail from "./FileItemDetail";
 
@@ -9,16 +10,24 @@ type Props = {
   loading: boolean;
   showDetail: boolean;
   setShowDetail: Dispatch<SetStateAction<boolean>>;
+  onFileUpdated?: (file: File) => void;
 };
-export default function FileListItem({ file, loading, showDetail, setShowDetail }: Props): JSX.Element {
+export default function FileListItem({ file, loading, showDetail, setShowDetail, onFileUpdated }: Props): JSX.Element {
   return (
     <List.Item
       id={file.fullPath}
       title={file.attributes.title}
       subtitle={file.attributes.publisher ?? file.attributes.source}
       accessories={(file.attributes.tags || []).map((tag) => ({ text: tag }))}
-      icon={Icon.Link}
-      actions={<DetailsActions file={file} showDetail={showDetail} setShowDetail={setShowDetail} />}
+      icon={getFaviconIcon(file.attributes)}
+      actions={
+        <DetailsActions
+          file={file}
+          showDetail={showDetail}
+          setShowDetail={setShowDetail}
+          onFileUpdated={onFileUpdated}
+        />
+      }
       detail={<FileItemDetail file={file} loading={loading} />}
     />
   );